feat(#167): move memo error fix & split memoThread component
